Fix an X11 memory allocatiom bug
For the case where the XSHM extension is not found, we have
to allocate memory for the screen display buffer ourselves. The
total should be the width times the height times the size of a
pixel. It turns out X11 doesn't have a defined pixel type: it
just uses unsigned long. However the allocation code is using
the size of a uint32_t. This is actually too small, since a
long is 64-bits wide on a 64-bit system.

This has been fixed so that we allocate a buffer that's actually
large enough.

This was overlooked because the XSHM extension is almost always
present, so we don't usually hit the code that allocates the
pixel buffer directly.

+++ software/firmware/linuxdoom-1.10/i_video.c
@@ -885,7 +885,7 @@ void I_InitGraphics(void)
     				ZPixmap,
     				0,
     				(char*)malloc((X_width * X_height) *
-				    sizeof(uint32_t)),
+				    sizeof(unsigned long)),
     				X_width, X_height,
     				32, 0);
